ci(fuzzing): treat timeout as a successful workflow run (#5256)
If we hit the fuzzing timeout, it means no crashes happenened,
therefore it makes no sense to treat it as a workflow failure.

"timeout" returns an exit code 124 when this happens,
and by default the Github runners treat non-zero ones as a failure.

Change the logic to convert the exit code 124 to 0 (success)
and pass everything else (legitimate failures) through.

Closes #5244.

Signed-off-by: Alex T. <frozen.and.blue@gmail.com>